Seperated config
Seperated config, updated pathing, will seperate the rest till cliFunction.js is obsolete
